06_Mario_A3C/requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning Environment
gymnasium>=0.29.0
gym-super-mario-bros>=7.4.0

# Visualization and Monitoring
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Image Processing
opencv-python>=4.8.0
numpy<2.0.0

# NES Emulator
nes-py>=8.2.1

# Keyboard Control
pygame>=2.5.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
